requirements.txt
pandas==2.2.2
playwright==1.48.0
pyyaml==6.0.2
beautifulsoup4==4.12.3
lxml==5.3.0